color_mapper.py
"""
AI Color Mapper for E-commerce
Uses OpenAI CLIP to map multilingual product color descriptions to standardized codes

Author: Yugesh
Built: 2024-2025
Purpose: Multilingual color mapping with image verification
"""

import torch
from PIL import Image, ImageEnhance
from transformers import CLIPProcessor, CLIPModel
import aiohttp
import nest_asyncio
import asyncio
from io import BytesIO
import pandas as pd
import re
import time
import unicodedata
import logging

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(levelname)s - %(message)s"
)

nest_asyncio.apply()

# Configuration
FILE_PATH = '/content/drive/MyDrive/PY Components/Color Map/Colour_Map_Template.xlsx'
OUTPUT_FILE = '/content/drive/MyDrive/PY Components/Color Map/Mapped_Colors_Output.xlsx'
CONFIDENCE_THRESHOLD = 0.7
IMAGE_SIZE = (1000, 1000)

# Initialize device and model
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
logging.info(f"Using device: {device}")

model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32").to(device)
processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32")

logging.info("✅ Color Mapper initialized - Created by Yugesh")
logging.info("📧 For support: Check GitHub repository")

# [Rest of implementation - see original script]
