requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
aiohttp>=3.9.0
pandas>=2.0.0
openpyxl>=3.1.0
nest-asyncio>=1.5.0
